+++ src/main/java/gov/inl/igcapt/components/AggregationDialog.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.swing.DefaultComboBoxModel;
 import javax.swing.table.DefaultTableModel;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
+import gov.inl.igcapt.view.IGCAPTgui;
 
 /**
  *(c) 2018 BATTELLE ENERGY ALLIANCE, LLC
