+++ src/main/java/gov/inl/igcapt/components/Heatmap.java
@@ -10,7 +10,7 @@ import static java.lang.Math.sqrt;
 import java.util.ArrayList;
 import java.util.List;
 import org.openstreetmap.gui.jmapviewer.Coordinate;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
+import gov.inl.igcapt.view.IGCAPTgui;
 import org.openstreetmap.gui.jmapviewer.JMapViewer;
 import org.openstreetmap.gui.jmapviewer.MapPolygonImpl;
 import gov.inl.igcapt.graph.SgNode;
