+++ src/main/java/gov/inl/igcapt/controllers/IGCAPTGraphMousePlugin.java
@@ -1,5 +1,6 @@
 package gov.inl.igcapt.controllers;
 
+import gov.inl.igcapt.view.IGCAPTgui;
 import edu.uci.ics.jung.algorithms.layout.GraphElementAccessor;
 import edu.uci.ics.jung.graph.Graph;
 import edu.uci.ics.jung.visualization.LayeredIcon;
