+++ src/main/java/gov/inl/igcapt/graph/SGEditingModalGraphMouse.java
@@ -39,7 +39,7 @@ import edu.uci.ics.jung.visualization.control.PickingGraphMousePlugin;
 import edu.uci.ics.jung.visualization.control.RotatingGraphMousePlugin;
 import edu.uci.ics.jung.visualization.control.ScalingGraphMousePlugin;
 import edu.uci.ics.jung.visualization.control.ShearingGraphMousePlugin;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
+import gov.inl.igcapt.view.IGCAPTgui;
 
 public class SGEditingModalGraphMouse<V, E> extends AbstractModalGraphMouse
         implements ModalGraphMouse, ItemSelectable {
