+++ src/main/java/gov/inl/igcapt/graph/SGEditingPopupGraphMousePlugin1.java
@@ -22,7 +22,7 @@ import edu.uci.ics.jung.graph.Graph;
 import edu.uci.ics.jung.visualization.VisualizationViewer;
 import edu.uci.ics.jung.visualization.picking.PickedState;
 import gov.inl.igcapt.components.DataModels.SgComponentData;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
+import gov.inl.igcapt.view.IGCAPTgui;
 
 import java.util.ArrayList;
 import java.util.List;
