+++ src/main/java/gov/inl/igcapt/view/AddClearAnalysisResultsMenuItem.java
@@ -1,7 +1,6 @@
 package gov.inl.igcapt.view;
 
 import gov.inl.igcapt.graph.GraphManager;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
 
 import javax.swing.*;
