+++ src/main/java/gov/inl/igcapt/view/AddShowHeatmapMenuItem.java
@@ -1,7 +1,6 @@
 package gov.inl.igcapt.view;
 
 import gov.inl.igcapt.components.HeatMapDialog;
-import org.openstreetmap.gui.jmapviewer.IGCAPTgui;
 
 import javax.swing.*;
