+++ src/main/java/org/openstreetmap/gui/jmapviewer/JMapViewer.java
@@ -1,6 +1,7 @@
 // License: GPL. For details, see Readme.txt file.
 package org.openstreetmap.gui.jmapviewer;
 
+import gov.inl.igcapt.view.IGCAPTgui;
 import edu.uci.ics.jung.algorithms.layout.AbstractLayout;
 import edu.uci.ics.jung.graph.Graph;
 import edu.uci.ics.jung.visualization.picking.PickedState;
